# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Auto Ledger API
  version: 0.1.0
  description: API for Auto Ledger web, mobile, and desktop clients.
security: []
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Log in a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: borys.test@example.com
              password: password1
      responses:
        "200":
          description: User logged in successfully
          headers:
            Set-Cookie:
              description: HTTP-only refresh token cookie
              schema:
                type: string
              example: "refreshToken=<opaque-token>; Expires=Thu, 22 Oct 2026 20:32:38 GMT; Path=/api/auth; HttpOnly; SameSite=Lax"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid login data or malformed JSON
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationErrorResponse"
                  - $ref: "#/components/schemas/ApiErrorResponse"
              examples:
                validationError:
                  summary: Invalid login fields
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid login data
                      fields:
                        email:
                          - Invalid email address
                invalidJson:
                  summary: Malformed JSON body
                  value:
                    error:
                      code: INVALID_JSON
                      message: Invalid JSON request body
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
        "500":
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: Internal server error
...
components:
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
          writeOnly: true
    LoginResponse:
      type: object
      additionalProperties: false
      required:
        - user
        - accessToken
      properties:
        user:
          $ref: "#/components/schemas/PublicUser"
        accessToken:
          type: string
...
    PublicUser:
      type: object
      additionalProperties: false
      required:
        - id
        - firstName
        - lastName
        - email
      properties:
        id:
          type: string
          pattern: "^[a-fA-F0-9]{24}$"
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
    ApiError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
            - fields
          properties:
            code:
              type: string
              const: VALIDATION_ERROR
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
  securitySchemes:
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refreshToken